apollo_state tests better organized
moved all tests into the tests dir which is per react dev convention
