+++ src/__tests__/Main.spec.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import MockTheme from '../testComponents/MockTheme';
-import Main from './Main';
+import Main from '../Main';
 // Apollo requirements
 import { ApolloProvider } from '@apollo/react-hooks';
 import { createMockClient } from 'mock-apollo-client';
 import { gql } from 'apollo-boost';
 import * as queries from '../graphql/queries';
-import me from '../__tests__/data';
+import me from './data';
 
 const meQuery = gql(queries.users.getMe.graphql);
